refactor(core): ♻️ move page heading to components
