+++ resources/skins.citizen.styles/components/StickyHeader.less
@@ -27,7 +27,7 @@
 		}
 	}
 
-	.page-heading {
+	.citizen-page-heading {
 		position: relative;
 		min-width: 0;
 	}
