+++ resources/skins.citizen.styles/layout.less
@@ -37,7 +37,7 @@
 	padding-top: var( --space-md );
 }
 
-.page-heading {
+.citizen-page-heading {
 	flex-grow: 1;
 }
